Moved files to /src/server
Moved SecurityCamera and Gate from the mock folder to the server folder as well as updated any necessary references.
