+++ src/server/Gate.java
@@ -1,4 +1,4 @@
-package mock;
+package server;
 
 import interfaces.GateInterface;
