+++ src/server/SecurityCamera.java
@@ -1,8 +1,9 @@
-package mock;
+package server;
 
 import javax.swing.ImageIcon;
 
 import interfaces.SecurityCameraInterface;
+import mock.Garage;
 
 public class SecurityCamera implements SecurityCameraInterface {
 	private static int count = 0;
