+++ src/server/Ticket.java
@@ -1,6 +1,5 @@
 package server;
 
-import mock.*;
 import interfaces.TicketInterface;
 
 import java.util.Date;
